+++ apps/mqtt_rest/src/mqtt_rest.erl
@@ -20,7 +20,7 @@
 -export([stop/1]).
 
 start(_Type, _Args) ->
-	Port = application:get_env(mqtt_server, port_rest, 8080),
+	Port = application:get_env(mqtt_rest, port, 8080),
 	lager:info([{endtype, server}], "Start mqtt_rest app = ~p Port:~p.~n", [_Args, Port]),	
 	mqtt_rest_server:start(mqtt_rest, #{ip=>{127,0,0,1}, port=>Port, net_opts=>[]}).