+++ apps/mqtt_rest/src/mqtt_rest_user_handler.erl
@@ -46,7 +46,7 @@
 
 init(Req, {Operations, LogicHandler, ValidatorMod}) ->
 	Storage =
-	case application:get_env(mqtt_rest, storage, dets) of
+	case application:get_env(mqtt_common, storage, dets) of
 		mysql -> mqtt_mysql_storage;
 		dets -> mqtt_dets_storage;
 		mnesia -> mqtt_mnesia_storage