+++ deployment_scripts/start_mqtt_server.sh
@@ -5,8 +5,8 @@
 erl \
  -pa _build/default/lib/*/ebin \
  -boot start_sasl \
- -config config-dev/sys \
- -sname mqtt-server@localhost \
+ -config config/sys_mqtt \
+ -sname mqtt_server@localhost \
  -s ssl \
  -eval "application:ensure_all_started(mqtt_server)" \
  -setcookie 'mqtt'