+++ start_stop_cluster_node_0.sh
@@ -1,10 +1,20 @@
 #!/bin/sh
 
 export PATH="$PATH:/usr/bin:/usr/local/bin:/usr/local/Cellar/erlang/28.0.2_1/bin"
+export RELX_REPLACE_OS_VARS=true
+export NODE_NAME=mqtt_server_0
+export PORT_REST=8080 \
+export PORT_CLEAR=18883 \
+export PORT_TLS=18483 \
+export PORT_WS=8880 \
+export PORT_WSS=4443 \
+export CLUSTER_NODES="'mqtt_server_0@MacBook-Pro','mqtt_server_1@MacBook-Pro'" \
+export MNESIA_MASTER=false \
+export MNESIA_DIR="'/opt/mqtt/cluster/node_0/mnesia'" 
+echo "Script to start/stop node $NODE_NAME of cluster"
 echo "arguments: $1 $2"
 REBAR3="/opt/local/bin/rebar3"
 $REBAR3 do version
-
 
 case "$1" in
 	dev)