+++ HISTORY.md
@@ -12,4 +12,11 @@ see for documentation: https://mqtt-server.sourceforge.io/
    Start up script introduced.
    
    Configuration is added.
- 
+
+#version 2.1.0
+   Restful http server is added.
+   
+   Support for MQTT protokol version 5.0 is added.
+
+#version 2.2.0
+   Mnesia storage is added. Mnesia is needed to create cluster of distributed MQTT servers.